+++ source/define/constant/ASYNC.ts
@@ -1,0 +1,4 @@
+/**
+  It is used to define asynchronous.
+*/
+export const ASYNC: string = 'async';