+++ source/define/constant/ASYNC_FUNCTION.ts
@@ -1,0 +1,4 @@
+/**
+  It expresses asynchronous functions.
+*/
+export const ASYNC_FUNCTION: string = 'AsyncFunction';