+++ source/define/constant/CLASS.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the "class" expression in object-oriented programming.
+*/
+export const CLASS: string = 'class';