+++ source/define/constant/MAP.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the "Map" data structure.
+*/
+export const MAP: string = 'map';