+++ source/define/constant/MAX_SAFE_INTEGER.ts
@@ -1,0 +1,5 @@
+/**
+  It represents the maximum safe integer.
+*/
+export const MAX_SAFE_INTEGER: number =
+  Number.MAX_SAFE_INTEGER || Math.pow(2, 53) - 1;