+++ source/define/constant/NAN.ts
@@ -1,0 +1,4 @@
+/**
+  It represents a non-numeric value.
+*/
+export const NAN: string = 'NaN';