+++ source/define/constant/NULL.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the null value.
+*/
+export const NULL: string = 'null';