+++ source/define/constant/OBJECT.ts
@@ -1,0 +1,4 @@
+/**
+  It represents objects.
+*/
+export const OBJECT: string = 'object';