+++ source/define/constant/PROMISE.ts
@@ -1,0 +1,4 @@
+/**
+  It represents asynchronous operations.
+*/
+export const PROMISE: string = 'promise';