+++ source/define/constant/SET.ts
@@ -1,0 +1,4 @@
+/**
+  It represents the data structure for storing unique values.
+*/
+export const SET: string = 'set';