+++ source/define/constant/SYMBOL.ts
@@ -1,0 +1,4 @@
+/**
+  It represents unique values.
+*/
+export const SYMBOL: string = 'symbol';