+++ source/define/index.ts
@@ -5,8 +5,19 @@
   ------------------------------------------------------------------------------
 */
 
+export { ASYNC_FUNCTION } from './constant/ASYNC_FUNCTION';
+export { ASYNC } from './constant/ASYNC';
 export { BOOLEAN } from './constant/BOOLEAN';
+export { CLASS } from './constant/CLASS';
 export { FUNCTION } from './constant/FUNCTION';
+export { MAP } from './constant/MAP';
+export { MAX_SAFE_INTEGER } from './constant/MAX_SAFE_INTEGER';
+export { NAN } from './constant/NAN';
+export { NULL } from './constant/NULL';
 export { NUMBER } from './constant/NUMBER';
+export { OBJECT } from './constant/OBJECT';
+export { PROMISE } from './constant/PROMISE';
+export { SET } from './constant/SET';
 export { STRING } from './constant/STRING';
+export { SYMBOL } from './constant/SYMBOL';
 export { UNDEFINED } from './constant/UNDEFINED';