+++ source/global.d.ts
@@ -1,3 +1,23 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/ban-types */
+/* eslint-disable @typescript-eslint/no-unnecessary-type-constraint */
+
+/**
+  Alias to create a function.
+
+  @template R - The type of parameters.
+  @template N - The return type.
+*/
+type Arrow<R extends List = any, N extends any = any> = (...args: R) => N;
+
+/**
+  Alias to create a async function.
+
+  @template R - The type of parameters.
+  @template N - The return type.
+*/
+type AsyncArrow<T extends List = any[], N = any> = Arrow<T, Promise<N>>;
+
 /**
   Type that checks if type `R` extends type `N`, and returns type `R` if true,
   otherwise returns `never`.
@@ -6,3 +26,9 @@
   @template N - The type to check against.
 */
 type Include<R, N> = R extends N ? R : never;
+
+/**
+  Alias representing a readonly array.
+  @template R - The type of elements in the list.
+*/
+type List<R = any> = ReadonlyArray<R>;