+++ source/type/collection/List.ts
@@ -1,0 +1,6 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/**
+  Alias representing a readonly array.
+  @template R - The type of elements in the list.
+*/
+export type List<R = any> = ReadonlyArray<R>;