+++ source/type/index.ts
@@ -6,4 +6,7 @@
   ------------------------------------------------------------------------------
 */
 
+export type { Arrow } from './collection/Arrow';
+export type { AsyncArrow } from './collection/AsyncArrow';
 export type { Include } from './collection/Include';
+export type { List } from './collection/List';